Fixed issues with display - turns out I should stick to even sized sprites
Sprites are positions using their center positions - so sprites with centers that are odd cause the sprite to be positioned at a .5 offset from the pixels
this causes awful distortion of pixels!
All of my sprites were odd numbers - probably because it is easier to make pixel art with odd numbers for circles?
Now the game looks 1000x better with all pixels being perfect!

Note: I didnt have to update the heart sprite - maybe because this is being positioned by the flexlayout?